Delete .babelrc
Babel plugins and presets in package.json so .babelrc can bet omitted. The .babelrc may cause problems with React Native for it contains it's own Babel configuration. Removing the .babelrc file and placing the Babel plugins and presets directly in the babel command makes it possible to use this library with React Native